requirements.txt
psycopg2-binary==2.9.9
redshift-connector==2.1.0
pandas==2.1.4
python-dotenv==1.0.0
tqdm==4.66.1
requests==2.31.0
slack-sdk>=3.12,<4

